refactor: split main.rs into model/cli/extractors/store/query modules
Pure module split with no behavior change: lib.rs + thin main.rs entry, shared
types in model.rs, per-language extractors, store (db/schema/sync), and query
(search/graph/context/stats). Tests relocated beside their code. Brigade smoke
unchanged (files=240 symbols=4313 edges=32164 imports=1669).
