+++ engines/code-graph/src/query/mod.rs
@@ -1,0 +1,11 @@
+//! Read-only query layer over the graph database.
+
+pub mod context;
+pub mod graph;
+pub mod search;
+pub mod stats;
+
+pub use context::build_context_pack;
+pub use graph::graph_edges;
+pub use search::search_symbols;
+pub use stats::stats;
