+++ engines/code-graph/src/store/schema.rs
@@ -1,0 +1,67 @@
+//! SQLite schema definition.
+
+use anyhow::Result;
+use rusqlite::Connection;
+
+/// Bumped when the on-disk schema changes; surfaced in JSON packs from Phase 2 on.
+pub const SCHEMA_VERSION: u32 = 1;
+
+pub fn init_schema(conn: &Connection) -> Result<()> {
+    conn.execute_batch(
+        r#"
+        CREATE TABLE IF NOT EXISTS files (
+            path TEXT PRIMARY KEY,
+            content_hash TEXT NOT NULL,
+            size INTEGER NOT NULL,
+            modified_at INTEGER NOT NULL,
+            indexed_at INTEGER NOT NULL,
+            language TEXT NOT NULL
+        );
+
+        CREATE TABLE IF NOT EXISTS symbols (
+            id TEXT PRIMARY KEY,
+            kind TEXT NOT NULL,
+            name TEXT NOT NULL,
+            qualified_name TEXT NOT NULL,
+            file_path TEXT NOT NULL,
+            start_line INTEGER NOT NULL,
+            end_line INTEGER NOT NULL,
+            signature TEXT NOT NULL,
+            container TEXT,
+            content_hash TEXT NOT NULL,
+            FOREIGN KEY(file_path) REFERENCES files(path) ON DELETE CASCADE
+        );
+
+        CREATE TABLE IF NOT EXISTS edges (
+            source TEXT NOT NULL,
+            target TEXT NOT NULL,
+            kind TEXT NOT NULL,
+            line INTEGER,
+            PRIMARY KEY(source, target, kind, line),
+            FOREIGN KEY(source) REFERENCES symbols(id) ON DELETE CASCADE,
+            FOREIGN KEY(target) REFERENCES symbols(id) ON DELETE CASCADE
+        );
+
+        CREATE TABLE IF NOT EXISTS imports (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            file_path TEXT NOT NULL,
+            module TEXT NOT NULL,
+            line INTEGER NOT NULL
+        );
+
+        CREATE VIRTUAL TABLE IF NOT EXISTS symbols_fts USING fts5(
+            symbol_id UNINDEXED,
+            name,
+            qualified_name,
+            signature,
+            file_path
+        );
+
+        CREATE INDEX IF NOT EXISTS idx_symbols_name ON symbols(name);
+        CREATE INDEX IF NOT EXISTS idx_symbols_file ON symbols(file_path);
+        CREATE INDEX IF NOT EXISTS idx_edges_source ON edges(source);
+        CREATE INDEX IF NOT EXISTS idx_edges_target ON edges(target);
+        "#,
+    )?;
+    Ok(())
+}
